@@ -5,32 +5,35 @@
   "dependencies": {
     "@babel/helper-builder-react-jsx": "^7.12.13",
     "@babel/helper-regex": "^7.10.5",
-    "@material-ui/core": "^4.11.4",
-    "@material-ui/lab": "^4.0.0-alpha.58",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@mui/lab": "^5.0.0-alpha.62",
+    "@mui/material": "^5.2.6",
+    "@mui/styles": "^5.2.3",
     "@types/node": "^15.6.1",
-    "@types/react-dom": "^17.0.5",
+    "@types/react-dom": "^17.0.11",
     "i18next": "^20.3.1",
     "i18next-browser-languagedetector": "^6.1.1",
-    "mdi-material-ui": "^6.22.1",
+    "mdi-material-ui": "^7.1.0",
     "node-sass": "^6.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-i18next": "^11.10.0",
     "react-scripts": "^4.0.3",
-    "typescript": "^4.3.2"
+    "typescript": "^4.5.4"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "netlify dev",
+    "start:app": "react-scripts start",
     "build": "run-p build:**",
     "build:app": "react-scripts build",
-    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": "react-app"
   },
   "devDependencies": {
-    "@types/react": "^17.0.8",
+    "@types/react": "^17.0.38",
     "npm-run-all": "^4.1.5"
   },
   "browserslist": {
