@@ -1,4 +1,4 @@
-import { Box } from "@material-ui/core";
+import { Box } from "@mui/material";
 import React from "react";
 import MCPBadge from "../../images/MCP_Badge.png";
 import CategoryHeader from "../CategoryHeader";
