@@ -1,5 +1,5 @@
 import React from "react";
-import { Box, IconButton, Link, Tooltip } from "@material-ui/core";
+import { Box, IconButton, Link, Tooltip } from "@mui/material";
 import { Github, Gitlab, StackOverflow } from "mdi-material-ui";
 import { useTranslation } from "react-i18next";
 
@@ -10,21 +10,21 @@ export default function ProfileLinks() {
     <Box>
       <Tooltip title={t("github").toString()}>
         <Link href="https://github.com/DanielBretzigheimer">
-          <IconButton>
+          <IconButton size="large">
             <Github />
           </IconButton>
         </Link>
       </Tooltip>
       <Tooltip title={t("gitlab").toString()}>
         <Link href="https://gitlab.com/dbretzigheimer95">
-          <IconButton>
+          <IconButton size="large">
             <Gitlab />
           </IconButton>
         </Link>
       </Tooltip>
       <Tooltip title={t("stack-overflow").toString()}>
         <Link href="https://stackoverflow.com/users/2119579/daniel-bretzigheimer">
-          <IconButton>
+          <IconButton size="large">
             <StackOverflow />
           </IconButton>
         </Link>
