@@ -1,9 +1,9 @@
-import { TimelineConnector, TimelineDot, TimelineSeparator } from "@material-ui/lab";
+import { TimelineConnector, TimelineDot, TimelineSeparator } from "@mui/lab";
 import React from "react";
 import { TimelineItemContent } from "../../model/TimelineItemContent";
 import EducationIcon from "mdi-material-ui/SchoolOutline";
 import ProjectIcon from "mdi-material-ui/StarOutline";
-import { Tooltip } from "@material-ui/core";
+import { Tooltip } from "@mui/material";
 
 interface SeperatorProps {
   isLast: boolean;
