@@ -1,5 +1,5 @@
 import React from "react";
-import { Link } from "@material-ui/core";
+import { Link } from "@mui/material";
 import { TimelineItemContent } from "../model/TimelineItemContent";
 import {
   skillAndroidOperationSystem as skillAndroidOperatingSystem,
